# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "conexionPos",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/v1/billingResolution": {
      "get": {
        "description": "get all billingResolutions of the BD",
        "tags": [
          "billingResolution"
        ],
        "responses": {
          "200": {
            "description": "get all billingResolutions successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }

      },
      "post": {
        "description": "insert a new billingResolutions in the BD",
        "tags": [
          "billingResolution"
        ],
        "requestBody": {
          "description": "billingResolutions create body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/billingResolutionsSchema"
              },
              "example": {
                "resolutionClass": "class prueba",
                "resolutionType": "type prueba",
                "resolutionNumber": 1234,
                "prefix": "res",
                "initialNumber": 2,
                "finalNumber": 3,
                "localBilling": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "billingResolution add successfully"
          },
          "401": {
            "description": "you can't create billingResolutions"
          }
        }
      }
    },
    "/v1/billingResolution/{id}": {
      "get": {
        "description": "get one billingResolution of the BD",
        "tags": [
          "billingResolution"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the billingResolution",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "get all billingResolution successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }

      },
      "put": {
        "description": "update one billingResolution of the BD",
        "tags": [
          "billingResolution"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the billingResolution",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "billingResolutions updated body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/billingResolutionSchema"
              },
              "example": {
                "resolutionClass": "class prueba mod",
                "resolutionType": "type prueba mod",
                "resolutionNumber": 123333,
                "prefix": "resmod",
                "initialNumber": 4,
                "finalNumber": 4,
                "localBilling": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "update billingResolution successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }
      },
      "delete": {
        "description": "delete one billingResolution of the BD",
        "tags": [
          "billingResolution"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the billingResolution",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "delete billingResolution successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }

      }
    },
    "/v1/banks":{
      "get": {
        "description": "get all banks of the BD",
        "tags": [
          "bank"
        ],
        "responses": {
          "200": {
            "description": "get all bank successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }

      },
      "post": {
        "description": "insert a new bank in the BD",
        "tags": [
          "bank"
        ],
        "requestBody": {
          "description": "banks create body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/bankSchema"
              },
              "example": {
                "name": "bancolombia",
                "accountingAccount": 123
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "bank add successfully"
          },
          "401": {
            "description": "you can't create bank"
          }
        }
      }
    },
    "/v1/banks/{id}": {
      "get": {
        "description": "get one bank of the BD",
        "tags": [
          "bank"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the bank",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "get all bank successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }

      },
      "put": {
        "description": "update one bank of the BD",
        "tags": [
          "bank"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the bank",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "banks updated body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/bankSchema"
              },
              "example": {
                "name": "bancolombia mod",
                "accountingAccount": 1212
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "update bank successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }
      },
      "delete": {
        "description": "delete one bank of the BD",
        "tags": [
          "bank"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the bank",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "delete bank successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }

      }
    },
...
  },
...
}